Rename gdk-window-hack -> gtk-priv-access
